Debugged version of the one trying to solve lock contention.
This can run but still can't solved the problem of lock contention.
Although I tried to improve lock granularity to decrease the common resources of bp__bulk_set(writer),
there are other issues need to be considered.
The execution time of bench-bulk rises and test-threaded appears deadlock between the processes.